Move projectloader to testutil
